requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cssselect2==0.7.0
Django==5.0.3
django-anymail==10.3
django-environ==0.11.2
djangorestframework==3.15.1
fonttools==4.50.0
html5lib==1.1
idna==3.6
pillow==10.2.0
pycparser==2.22
pydyf==0.9.0
pyphen==0.14.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.10.0
urllib3==2.2.1
weasyprint==61.2
webencodings==0.5.1
zopfli==0.2.3
